lastet

+++ NuSysApp/Components/Workspace/WorkspaceView.xaml.cs
@@ -13,6 +13,7 @@ using Windows.UI;
 using Windows.ApplicationModel.Core;
 using Windows.UI.Core;
 using System.Diagnostics;
+using Windows.UI.Input.Inking;
 using Windows.UI.Xaml.Media.Animation;
 
 // The Blank Page item template is documented at http://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
@@ -55,26 +56,28 @@ namespace NuSysApp
                     case GestureRecognizer.GestureType.None:
                         break;
                     case GestureRecognizer.GestureType.Scribble:
-                        vm.CheckForInkNodeIntersection(model);
-                        model.Delete();
+                        var deletedSome = vm.CheckForInkNodeIntersection(model);
+                        if (deletedSome)
+                            model.Delete();
                         break;
                 }
 
-                await CheckForGroupCreation(model);
+                await CheckForTagCreation(model);
             };
         }
 
-        private async Task<bool> CheckForGroupCreation(InqLineModel line)
+        private async Task<bool> CheckForTagCreation(InqLineModel line)
         {
             var wsmodel = (DataContext as WorkspaceViewModel).Model as WorkspaceModel;
             var Model = wsmodel.InqCanvas;
             var outerRect = Geometry.PointCollecionToBoundingRect(line.Points.ToList());
 
-            if (outerRect.Width * outerRect.Height < 100 * 100)
+            if (outerRect.Width * outerRect.Height < (100.0 / Constants.MaxCanvasSize * 100.0 /Constants.MaxCanvasSize) )
                 return false;
 
-            var idsToDelete = new List<string>();
+            var idsToDelete = new List<InqLineModel>();
             var encompassedLines = new List<InqLineModel>();
+
             foreach (var otherLine in Model.Lines.Where(l => l != line))
             {
                 var innerRect = Geometry.PointCollecionToBoundingRect(otherLine.Points.ToList());
@@ -83,7 +86,7 @@ namespace NuSysApp
                 if (Math.Abs(innerRect2.Width - innerRect.Width) < 70 && Math.Abs(innerRect2.Height - innerRect.Height) < 70)
                 {
 
-                    idsToDelete.Add(otherLine.Id);
+                    idsToDelete.Add(otherLine);
                     InqLineModel newModel = new InqLineModel(DateTime.UtcNow.Ticks.ToString())
                     {
                         Stroke = otherLine.Stroke,
@@ -106,17 +109,34 @@ namespace NuSysApp
                 return false;
             }
 
-            /*
+            
             foreach (var idToDelete in idsToDelete)
             {
-                NetworkConnector.Instance.RequestDeleteSendable(idToDelete);
+               idToDelete.Delete();
             }
            
-
-
             var title = await InkToText(encompassedLines);
             var dict = new Dictionary<string, string>();
             dict["title"] = title;
+           
+            var tagNodePos = new Point(outerRect.X + outerRect.Width / 6, outerRect.Y + outerRect.Height / 6);
+
+            var m = new Message();
+            m["x"] = tagNodePos.X * Constants.MaxCanvasSize;
+            m["y"] = tagNodePos.Y * Constants.MaxCanvasSize;
+            m["width"] = 400;
+            m["title"] = title;
+            m["height"] = 400;
+            m["nodeType"] = NodeType.Tag.ToString();
+            m["autoCreate"] = true;
+            m["creators"] = new List<string>() { SessionController.Instance.ActiveWorkspace.Id };
+
+
+            await SessionController.Instance.NuSysNetworkSession.ExecuteRequest(new NewNodeRequest(m));
+
+            /*
+            await NetworkConnector.Instance.RequestNewGroupTag(tagNodePos.X.ToString(), tagNodePos.Y.ToString(), title, dict, addCallback);
+
             Action<string> addCallback = delegate (string s)
             {
                 NetworkConnector.Instance.RequestDeleteSendable(line.Id);
@@ -136,10 +156,9 @@ namespace NuSysApp
                     }
                 }
             };
-            var tagNodePos = new Point(outerRect.X + outerRect.Width / 6, outerRect.Y + outerRect.Height / 6);
-            await NetworkConnector.Instance.RequestNewGroupTag(tagNodePos.X.ToString(), tagNodePos.Y.ToString(), title, dict, addCallback);
+            */
             return true;
-             */
+             
             return false;
         }
 
@@ -176,7 +195,8 @@ namespace NuSysApp
                 case Options.SelectNode:
                     var nodeManipulationMode = new NodeManipulationMode(this);
                     await
-                        SetViewMode(new MultiMode(this, nodeManipulationMode, new DuplicateNodeMode(this), new PanZoomMode(this), new SelectMode(this),
+                        SetViewMode(new MultiMode(this, nodeManipulationMode, new DuplicateNodeMode(this),
+                            new PanZoomMode(this), new SelectMode(this), new TagNodeMode(this),
                             new FloatingMenuMode(this), new CreateGroupMode(this, nodeManipulationMode)));
                     break;
                 case Options.SelectMarquee:
@@ -189,19 +209,23 @@ namespace NuSysApp
                     break;
                 case Options.AddTextNode:
                     await
-                        SetViewMode(new MultiMode(this, new AddNodeMode(this, NodeType.Text, isFixed), new FloatingMenuMode(this)));
+                        SetViewMode(new MultiMode(this, new AddNodeMode(this, NodeType.Text, isFixed),
+                            new FloatingMenuMode(this)));
                     break;
                 case Options.AddWeb:
                     await
-                        SetViewMode(new MultiMode(this, new AddNodeMode(this, NodeType.Web, isFixed), new FloatingMenuMode(this)));
+                        SetViewMode(new MultiMode(this, new AddNodeMode(this, NodeType.Web, isFixed),
+                            new FloatingMenuMode(this)));
                     break;
                 case Options.AddAudioCapture:
                     await
-                        SetViewMode(new MultiMode(this, new AddNodeMode(this, NodeType.Audio, isFixed),new FloatingMenuMode(this)));
+                        SetViewMode(new MultiMode(this, new AddNodeMode(this, NodeType.Audio, isFixed),
+                            new FloatingMenuMode(this)));
                     break;
                 case Options.AddMedia:
                     await
-                        SetViewMode(new MultiMode(this, new SelectMode(this), new AddNodeMode(this, NodeType.Document, isFixed), new FloatingMenuMode(this)));
+                        SetViewMode(new MultiMode(this, new SelectMode(this),
+                            new AddNodeMode(this, NodeType.Document, isFixed), new FloatingMenuMode(this)));
                     break;
                 case Options.AddRecord:
                     var sessionView = SessionController.Instance.SessionView;
@@ -237,6 +261,72 @@ namespace NuSysApp
                             new FloatingMenuMode(this)));
                     break;
             }
+        }
+
+        public async Task<string> InkToText()
+        {
+            var wsmodel = (DataContext as WorkspaceViewModel).Model as WorkspaceModel;
+            var Model = wsmodel.InqCanvas;
+            if (Model.Lines == null || Model.Lines.Count == 0)
+                return string.Empty;
+
+            var im = new InkManager();
+
+            var b = new InkStrokeBuilder();
+
+            foreach (var inqLineModel in Model.Lines)
+            {
+                var pc = new PointCollection();
+                foreach (var point2D in inqLineModel.Points)
+                {
+                    pc.Add(new Point(point2D.X, point2D.Y));
+                }
+                var stroke = b.CreateStroke(pc);
+                im.AddStroke(stroke);
+            }
+
+            var result = await im.RecognizeAsync(InkRecognitionTarget.All);
+            return result[0].GetTextCandidates()[0];
+        }
+
+        public async Task<string> InkToText(List<InqLineModel> inqLineModels)
+        {
+            if (inqLineModels.Count == 0)
+                return string.Empty;
+
+            var im = new InkManager();
+
+
+            var b = new InkStrokeBuilder();
+
+            foreach (var inqLineModel in inqLineModels)
+            {
+                var pc = new PointCollection();
+                foreach (var point2D in inqLineModel.Points)
+                {
+                    pc.Add(new Point(point2D.X * Constants.MaxCanvasSize, point2D.Y * Constants.MaxCanvasSize));
+                }
+
+                var stroke = b.CreateStroke(pc);
+                im.AddStroke(stroke);
+            }
+
+            var result = await im.RecognizeAsync(InkRecognitionTarget.All);
+            var r = result[0].GetTextCandidates()[0];
+            if (r == "as")
+            {
+                r = "CSS";
+            }
+            if (r == "(55")
+            {
+                r = "CSS";
+            }
+            if (r == "Is")
+            {
+                r = "JS";
+            }
+            Debug.WriteLine(r);
+            return r;
         }
     }
 }
